move facebook sdk to services folder, change references
